added a submove unit and functions to do partial moves. then added a "general" move validation function wich uses sub moves instead of the distance deltas. If we can get the sub-move general check to validate all the genral moves then the output should be the same with the general or with the type specific checks.
Once we get that working, we can remove the specific functions.

Then we can start working on eliminating the "moves" themselves, and just replace them with submoves. then we can achieve all kinds of weaving without adding special cases. the submoves will do all the work.

this means we will have to come up with a "best sub move" function. This is harder than the reguler best move (we will need a state machine).