Big update to add plumbing for restarting threads. Now the kernel starts in an atomic critical section. Calling SchedulerStartThread must be done from a critical section.
Now we start up in a critical atomic section. This allows
us to call SchedulerCreateThread from startup and while
running.

This led to me fixing the startup system. Now we call
each startup routine strictly based upon the dependancy
graph. This resulted in me having to change the way
context starts up. Because it has to be done before the
scheduler we need to set ActiveStack to NULL for awile.
This breaks our old rule, but it seems to be ok.