Adding new tree data structure. Adding new tree tests.
Trees are different than the other link structures
(heaps included) we have offered so far because
they allow for any element to be deleted.

For now the AVL implementation is not a true AVL.
I just threw together a binary search tree. 
I'll get around to fixing up AVL later.